chore: move responsive functions to utils package
