+++ .changeset/proud-lemons-sell.md
@@ -1,0 +1,5 @@
+---
+"@chakra-ui/utils": minor
+---
+
+Move responsive utilities to utils to prevent duplicate work
